Deprecate `PhoneAuthProvider.MaxTimeoutMs`
This is a read-only valued used to set the upper bound for
`autoVerifyTimeOutMs`.  However, the value was set too low.
Android SDK has increase the range to 30 to 120 seconds, while iOS
SDK does not have a way to override timeout.

This is a change to deprecate this public property so that
the underlying SDK can determine the range for timeout.