Add workflow that builds tvOS (#484)
Fixed up workflow to build Unity for tvOS targets.

Workflow is triggered by manual triggers still - and not by the `Firebase Unity SDK build` which kicks-off builds for all of the platforms. This is simply so we don't release the tvOS libraries. Later, when the tvOS release is ready, we'll add this `Build tvOS (SubWorkflow)` to the main build list.