remove BuildTvos method from Firestore/Editor/Builder.cs (#541)
Remove an attempt to support tvOS for Firestore in the Editor/Builder.cs code. This failed to compile due to Unity not being able to resolve the tvOSSdkVersion enumerated type, for some reason.

However, it seems that these methods aren't invoked anywhere. Removing for now. Will add back and fix if Firestore fails to build for tvOS in future testing.